Code splited furthur, by the filter and the table displaying all tasks.
