+++ src/App.test.js
@@ -6,6 +6,6 @@ import App from './App';
 describe('App component', () => {
   render(<Provider store={store}><App /></Provider>);
   it('should render a component onto the screen', () => {
-      expect(true).toBeTruthy();
+    expect(true).toBeTruthy();
   });
 });
